+++ src/main/kotlin/registry/PlayerStatus.kt
@@ -1,6 +1,6 @@
 package dev.gangster.registry
 
 data class PlayerStatus(
-    val playerId: Long,
+    val playerId: Int,
     val onlineSince: Long,
 )
